Fixed how celery sqlalchemy models are used
Before I was doing a rather stupid thing and trying to creating my own models in sqlalchemy for the celery results backend. 

The much smarter solution(and more reliable) is just to use the models that are exposed directly by celery itself! Thus removes the need for rewriting and matching things to the underlying code of celery. Resulting in much simpler code that works better.